had to move components out of the pages directory to build the app
