+++ src/components/othello/GameBoard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../index.module.css';
+import styles from '../../pages/index.module.css';
 
 import { findCapturedTokenKeys, findAllPossibleMoves } from '../../utils/game';
 import type { GameState } from '../../utils/game';
