+++ src/components/othello/Info.tsx
