+++ src/pages/othello/index.tsx
@@ -4,8 +4,8 @@ import { type NextPage } from 'next';
 
 import styles from '../index.module.css';
 import { createInitialState } from '../../utils/game';
-import GameBoard from './GameBoard';
-import Info from './Info';
+import GameBoard from '../../components/othello/GameBoard';
+import Info from '../../components/othello/Info';
 
 const Othello: NextPage = () => {
   const [gameState, setGameState] = React.useState(createInitialState());
@@ -26,11 +26,11 @@ const Othello: NextPage = () => {
       <main className={styles.othellomain}>
         <h1>Othello</h1>
 
-        {hydrated && <Info gameState={gameState} />}
+        {hydrated && gameState ? <Info gameState={gameState} /> : null}
 
-        {hydrated && <GameBoard gameState={gameState} setGameState={setGameState} />}
+        {hydrated && gameState ? <GameBoard gameState={gameState} setGameState={setGameState} /> : null}
 
-        {hydrated && Object.values(gameState.board).every((cell) => !!cell.occupiedBy) ? (
+        {hydrated && gameState && Object.values(gameState.board).every((cell) => !!cell.occupiedBy) ? (
           <p>
             {Object.values(gameState.board).filter((cell) => cell.occupiedBy === 'white').length >
             Object.values(gameState.board).filter((cell) => cell.occupiedBy === 'black').length
